requirements.txt
flask
google-generativeai
python-dotenv
elevenlabs
base64
